services/agents/nepa-advisor/pyproject.toml
[project]
name = "nepa-advisor"
version = "1.0.0"
description = "RANGER NEPA Advisor Agent - AI-powered NEPA compliance guidance"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "google-adk>=0.1.0",
    "google-genai>=0.1.0",
    "agent-common",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
